frontend/backend/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
PyMySQL==1.1.0
Werkzeug==3.0.1
